// src/api/user/user.types.d.ts
export declare enum GenderType {
    Male = "Male",
    Female = "Female",
    Other = "Other"
}